refactor: Rename StandardSeeder module as Standard
So it is symmetric with the Bim module and both represent an edition.
