+++ app/seeders/standard/basic_data/activity_seeder.rb
@@ -25,27 +25,17 @@
 #
 # See COPYRIGHT and LICENSE files for more details.
 #++
-module StandardSeeder
+module Standard
   module BasicData
-    class PrioritySeeder < ::BasicData::PrioritySeeder
+    class ActivitySeeder < ::BasicData::ActivitySeeder
       def data
-        color_names = [
-          'cyan-1', # low
-          'blue-3', # normal
-          'yellow-7', # high
-          'grape-5' # immediate
-        ]
-
-        # When selecting for an array of values, implicit order is applied
-        # so we need to restore values by their name.
-        colors_by_name = Color.where(name: color_names).index_by(&:name)
-        colors = color_names.collect { |name| colors_by_name[name].id }
-
         [
-          { name: I18n.t(:default_priority_low),       color_id: colors[0], position: 1, is_default: false },
-          { name: I18n.t(:default_priority_normal),    color_id: colors[1], position: 2, is_default: true  },
-          { name: I18n.t(:default_priority_high),      color_id: colors[2], position: 3, is_default: false },
-          { name: I18n.t(:default_priority_immediate), color_id: colors[3], position: 4, is_default: false }
+          { name: I18n.t(:default_activity_management),    position: 1, is_default: true  },
+          { name: I18n.t(:default_activity_specification), position: 2, is_default: false },
+          { name: I18n.t(:default_activity_development),   position: 3, is_default: false },
+          { name: I18n.t(:default_activity_testing),       position: 4, is_default: false },
+          { name: I18n.t(:default_activity_support),       position: 5, is_default: false },
+          { name: I18n.t(:default_activity_other),         position: 6, is_default: false }
         ]
       end
     end
